+++ pylabrobot/centrifuge/__init__.py
@@ -1,2 +1,4 @@
-from .centrifuge import Centrifuge
-from .vspin import VSpin
+from .access2 import Access2
+from .centrifuge import Centrifuge, Loader
+from .standard import LoaderNoPlateError
+from .vspin import Access2Backend, VSpin